namespaces image load and uses throughout
